created Question class
