+++ src/entities/Question.java
@@ -1,12 +1,45 @@
 package entities;
 
-public interface Question {
+public class Question {
+    private final String questionText;
+    private final String correctAnswer;
+    private final String[] incorrectAnswers;
+    private final String difficulty;
+    private final String category;
+    private final String type;
 
-    String getQuestionText();
+    public Question(String questionText, String correctAnswer,
+                    String[] incorrectAnswers, String difficulty, String category, String type){
+        this.questionText = questionText;
+        this.correctAnswer = correctAnswer;
+        this.incorrectAnswers = incorrectAnswers;
+        this.difficulty = difficulty;
+        this.category = category;
+        this.type = type;
+    }
 
-    String getCorrectAnswer();
+    public String getQuestionText(){
+        return questionText;
+    }
 
-    String getDifficulty();
+    public String getCorrectAnswer(){
+        return correctAnswer;
+    }
 
-    String getCategory();
+    public String[] getIncorrectAnswers(){
+        return incorrectAnswers;
+    }
+
+    public String getDifficulty(){
+        return difficulty;
+    }
+
+    public String getCategory(){
+        return category;
+    }
+
+    public String getType(){
+        return type;
+    }
+
 }
